requirement.txt
# pip install -r requirement.txt
appium-python-client==1.2.0
pytest==5.4.1
allure-pytest==2.8.13
pytest-assume==2.2.1
pyYAML==5.3.1
jsonpath==0.82